// packages/frontend/e2e-smoke-tests/interfaces/tab-interfaces.ts
export enum ETabUrl {
    indicator = 'Indicator',
    instruments = 'Instruments',
    realAccounts = 'Real%20Accounts',
    virtualAccounts = 'Virtual%20Accounts',
    productLogs = 'Product%20Logs',
    orderBook = 'Order%20Book',
    position = 'Positions',
    balances = 'Balances',
    wsRequest = 'WS%20Request',
    config = 'Config',
    state = 'State',
    customView = 'Custom%20View',
    activeOrders = 'Active%20Orders',
    command = 'Command',
    addTask = 'Add%20Task',
    activeTasks = 'Active%20Tasks',
    archivedTasks = 'Archived%20Tasks',
    statusMessages = 'Status%20Messages',
    robotPositions = 'Robot%20Positions',
    robotBalances = 'Robot%20Balances',
    pnl = 'PNL',
    arb = 'ARB',
    trades = 'Trades',
    profits = 'Profits',
    arbVolume = 'ARB%20Volume',
    arbMaker = 'ARB%20Maker',
    arbTaker = 'ARB%20Taker',
    arbFees = 'ARB%20Fees',
    monthlyStats = 'Monthly%20Stats',
    tasks = 'Tasks',
    dashboards = 'Dashboards',
    runs = 'Runs',
    runsInfo = 'Runs%20Info',
    details = 'Task',
    configs = 'Run%20Configs',
    states = 'Run%20States',
    buildInfo = 'Run%20Build%20Info',
    indicators = 'Indicators',
    suggestedTransfers = 'Suggested%20Transfers',
    transfersHistory = 'Transfers%20History',
    coinTransferDetails = 'Coin%20Transfer%20Details',
    distribution = 'Distribution',
    gathering = 'Gathering',
    sendDataToAnalyse = 'Send%20data%20to%20analyse',
    manualTransfer = 'Manual%20Transfer',
    internalTransfers = 'Internal%20Transfers',
    internalTransfersHistory = 'Internal%20Transfers%20History',
    addTransferBlockingRule = 'Add%20Transfer%20Blocking%20Rule',
    transferBlockingRules = 'Transfer%20Blocking%20Rules',
    addAmountLimitsRule = 'Add%20Amount%20Limits%20Rule',
    amountLimitsRules = 'Amount%20Limits%20Rules',
    addAutoTransferRule = 'Add%20Auto%20Transfer%20Rule',
    autoTransferRules = 'Auto%20Transfer%20Rules',
    componentStatuses = 'Component%20Statuses',
    pumpAndDump = 'Pump%20%26%20Dump',
}